Rebuild tilemaps
Levels 1-3 now render properly! Please take a look at how the level is constructed- this is different than how it was done before: The collision layer is now the same layer as the "Walls" layer, and the Overlap layer consists of all the "tops" of the walls, regardless of position.